+++ R/DeeDeeExperiment-pkg.R
@@ -14,11 +14,14 @@
 #' @importFrom DESeq2 results DESeq vst
 #' @importFrom edgeR topTags
 #' @importFrom limma topTable
-#' @importFrom stats na.omit
+#' @importFrom stats na.omit setNames
+#' @importClassesFrom SingleCellExperiment SingleCellExperiment
 #' @importClassesFrom SummarizedExperiment RangedSummarizedExperiment
 #' @importFrom S4Vectors metadata metadata<- DataFrame SimpleList
-#' @importFrom SummarizedExperiment rowData rowData<- mcols assays
-#' rowData rowData<- SummarizedExperiment
+#' @importFrom SummarizedExperiment rowData rowData<- mcols assays colData
+#' colData<- assay
+#' @importFrom SingleCellExperiment int_colData int_colData<- int_metadata
+#' int_metadata<- SingleCellExperiment
 #' @importFrom utils packageVersion
 #' @importFrom methods show as callNextMethod is new validObject
 #' @importFrom cli cli_alert_success cli_alert_info